+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Each release is divided into the following main categories:
 - Updated text cropping docs [#103](https://github.com/allink/allink-core-static/pull/103)
 - Removed inverted variant for text selection [#104](https://github.com/allink/allink-core-static/pull/104)
 - Optimized scroll performance in triggerClassOnScroll [#110](https://github.com/allink/allink-core-static/pull/110)
+- Site title changes accordingly by opening a softpage [#116](https://github.com/allink/allink-core-static/pull/116)
 
 ### FIXES
 - Fixed radio / checkbox spacing [#105](https://github.com/allink/allink-core-static/pull/105)